+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/project/ProjectResponseService.java
@@ -14,7 +14,7 @@ import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.Service;
 
 /**
- * 资产目录响应服务。
+ * 工程响应服务。
  *
  * @author DwArFeng
  * @since 1.0.3