+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 升级部分依赖版本。
 
 - 新建 Controller。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.project.TaskController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.project.TaskTypeIndicatorController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.project.ProjectController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.project.PopController。